+++ teaser/include/teaser/macros.h
@@ -63,7 +63,7 @@
 #define TEASER_DEBUG_START_TIMING(s) auto t_start_##s = clock_##s.now();
 #define TEASER_DEBUG_STOP_TIMING(s)                                                                \
   auto t_end_##s = clock_##s.now();                                                                \
-  std::chrono::duration<double, std::milli> diff_dur_##s = t_end_##s - t_start_##s;                \
-  double diff_##s = diff_dur_##s.count();
-#define TEASER_DEBUG_GET_TIMING(s)(double)(diff_##s / 1000.0)
+  std::chrono::duration<float, std::milli> diff_dur_##s = t_end_##s - t_start_##s;                \
+  float diff_##s = diff_dur_##s.count();
+#define TEASER_DEBUG_GET_TIMING(s)(float)(diff_##s / 1000.0)
 #endif
